Wallet Integration: Include basic wallet connector for user.wallets
We add a new field `wallets` to our user's schema.  We intend to have an
authenticated endpoint(s) that can add/remove to this array for the
current user.

However, this change simply modifies the profile view and edit pages to:
1. Show any wallets (currently blank)
2. Add a wallet connector

NOTE: This requires us to load the `NftToolkit` from CDN as we rely on
the WildTangz dApp connector.

NOTE: There is an unrelated change for Mac OS X development (.DS_Store)

[ Documentation: None required (no setup changes) ]
[ Testing: npm run devStart ]