Wallet Connector: /remove_wallet endpoint for remove wallet
This commit adds a DELETE endpoint for the user that allows them to
remove a wallet.  The algorithm on the backend is simple:
1. Get the user from the current session
2. Check that the wallet is actually linked
3. Pull it from MongoDB using the `$pull` operator

Then on the frontend, we add JavaScript and styling to make rows with
the existing linked wallets and a trash icon next to each.

NOTE: We move the wallet loading to JavaScript to encourage code reuse,
though when a new wallet is successfully linked, though it's not
strictly necessary.

[ Testing: npm run devStart ]